requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
pyyaml==6.0.1
pycryptodome==3.19.0
